refactored the code for scraping
